refactor: Users domain folder structure
