--- src/domains/users/Auth/api/index.tsx
@@ -1,1 +1,0 @@
-export * as AuthService from "./AuthService";
