--- src/domains/users/Auth/features/index.tsx
@@ -1,3 +1,0 @@
-export * from "./LoginPage";
-export * from "./SignUpPage";
-export * from "./ProviderSignUpPage";
