--- src/domains/users/CustomerSignupPage/api/index.tsx
@@ -1,2 +1,0 @@
-export * from "./effects";
-export * from "./CustomerSignupReducer";
