--- src/domains/users/CustomerSignupPage/components/index.tsx
@@ -1,3 +1,0 @@
-export * from "./CustomerSignupForm";
-export * from "./CustomerSignupPage";
-export * from "./Notification";
