--- src/domains/users/CustomerSignupPage/index.tsx
@@ -1,3 +1,0 @@
-export * from "./components";
-export * from "./api";
-export * from "./types";
