+++ src/domains/users/Login/LoginPage.tsx
@@ -7,7 +7,7 @@ import { zodResolver } from "@hookform/resolvers/zod";
 import { Button, Checkbox, Form, FormLayout, Page } from "@shopify/polaris";
 import { routerPush } from "@Utils/router";
 import { TextField } from "@ComponentsV2/components/TextField";
-import { signIn } from "@Users/Auth/api/AuthService";
+import { signIn } from "../api";
 
 const formSchema = z.object({
   email: z
