+++ src/domains/users/ProviderSignUp/hooks/useProviderSignUpForm.tsx
@@ -2,8 +2,8 @@ import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
 import { z } from "zod";
 
-import { createCustomerUser } from "@Users/Auth/api/AuthService";
-import { routerPush } from "@Utils/index";
+import axios from "axios";
+import { createUser } from "@Users/api";
 
 export const formSchema = z
   .object({
@@ -23,12 +23,24 @@ export const formSchema = z
 
 export type SignUpFormState = z.infer<typeof formSchema>;
 
-export const submitHandler = async (data: SignUpFormState) => {
-  await createCustomerUser(data);
-  routerPush("/");
+const fetchProviderOnboardingUrl = async (userId: string) => {
+  const response = await axios.post(
+    `${process.env.NEXT_PUBLIC_API_URL}api/v1/providers`,
+    {
+      UserId: userId,
+    }
+  );
+
+  return response.data.data.accountLink.url;
 };
 
-export const useSignUpForm = () => {
+export const submitHandler = async (data: SignUpFormState) => {
+  const userId = await createUser(data);
+  const providerOnboardingUrl = await fetchProviderOnboardingUrl(userId);
+  window.location.assign(providerOnboardingUrl);
+};
+
+export const useProviderSignUpForm = () => {
   const { control, handleSubmit } = useForm<SignUpFormState>({
     mode: "onBlur",
     resolver: zodResolver(formSchema),
